Initial commit
Install reveal.js presentation framework using bower:

bower install reveal.js --save

Setup an initial slide show with a title page and resources page to conclude.

It seems like the bower_components directory does not belong in the repository,
but I am not familar with practices using the tool so it can be included for
now.